web/backend/requirements-essential.txt
# Dépendances essentielles pour Bubble Cleaner
# API & Framework
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
alembic==1.13.1

# Email
fastapi-mail==1.4.1
requests>=2.32.3

# Traitement d'images
opencv-python-headless==4.11.0.86
pillow==11.3.0
numpy<2.0
easyocr==1.7.0
pytesseract==0.3.13

# IA & ML
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
scikit-image==0.25.2
scipy==1.15.3
git+https://github.com/facebookresearch/detectron2.git

# Utilitaires
python-bidi==0.6.6
shapely==2.1.1

# OpenAI
openai>=1.0.0 